Release 2020.1

+++ dzVents/runtime/misc/testdzVents.sh
@@ -2,7 +2,7 @@ basedir=${1:-$test}
 #
 #
 # Test script for dzVents under domoticz.
-# tested an working on Raspberry Pi with Debian Stretch
+# tested and working on Raspberry Pi with Debian Buster
 # Call with base directory as parm
 #
 
@@ -32,7 +32,6 @@ if [ ! -f "$basedir/dzVents/runtime/integration-tests/testIntegration.lua" ]; th
 	exit 1
 fi
 
-
 NewVersion=$($basedir/domoticz --help | grep Giz | awk '{print $5}') 2>&1 >/dev/null
 export NewVersion
 if [[ $? -ne 0 ]];then
@@ -41,6 +40,14 @@ if [[ $? -ne 0 ]];then
 else
 	echo dzVents will be tested against domoticz version $NewVersion.
 fi
+ 
+currenttime=$(date +%H:%M)
+if [[ "$currenttime" > "00:00" ]] && [[ "$currenttime" < "00:31" ]]; then 
+	echo Script cannot be execute dbetween 00:00 and 00:30. Time checks will fail
+	exit 1
+fi
+
+dzVersion=$(grep DZVERSION dzVents/runtime/Utils.lua | head -1 | grep -oP "'.*'"  | sed s/\'//g)
 
 function leadingZero
 	{
@@ -73,7 +80,7 @@ checkStarted()
 		# echo $result $loopCounter $maxSeconds
 		if [[ $result -eq 1 && $loopCounter -le $maxSeconds ]];then
 			printf "%s" "."
-			(($loopCounter++))
+			loopCounter=$((loopCounter+1))
 		else
 			if [[ $result -eq 1 ]];then
 				 echo
@@ -88,18 +95,20 @@ checkStarted()
 
 function cleanup
 	{
-		ps -ef | grep domoticz
-		rm domoticz.log[0-9][0-9]*
-		find . -type f -name 'domoticz.db_*' -mmin +30 -exec rm {} \;
+		ps -aux | grep [f]domoticz
+		if [[ $? -eq 1 ]]; then
+			rm domoticz.log[0-9][0-9]*
+			find . -type f -name 'domoticz.db_*' -mmin +30 -exec rm {} \;
+		fi
 	}
 
 function stopBackgroundProcesses
 	{
-		pkill node 2>&1 >/dev/null
-		pkill domoticz 2>&1 >/dev/null
+		pkill node  2>&1 >/dev/null
+		pkill domoticz  2>&1 >/dev/null
 		if [[ $1 -eq 1 ]] ;then
 			exit $1 2>&1 >/dev/null
-		fi	
+		fi
 	}
 
 function fillTimes
@@ -107,34 +116,67 @@ function fillTimes
 		Device_ExpectedSeconds=10
 		Domoticz_ExpectedSeconds=10
 		EventHelpers_ExpectedSeconds=10
-		EventHelpersStorage_ExpectedSeconds=20
-		HTTPResponse_ExpectedSeconds=20
+		EventHelpersStorage_ExpectedSeconds=10
+		HTTPResponse_ExpectedSeconds=10
+		Lodash_ExpectedSeconds=10
 		ScriptdzVentsDispatching_ExpectedSeconds=10
 		TimedCommand_ExpectedSeconds=10
-		Time_ExpectedSeconds=10
+		Time_ExpectedSeconds=30
 		Utils_ExpectedSeconds=10
 		Variable_ExpectedSeconds=10
 		ContactDoorLockInvertedSwitch_ExpectedSeconds=30
 		DelayedVariableScene_ExpectedSeconds=80
-		EventState_ExpectedSeconds=180
-		Integration_ExpectedSeconds=300
+		EventState_ExpectedSeconds=190
+		Integration_ExpectedSeconds=330
 		SelectorSwitch_ExpectedSeconds=100
+		SystemAndCustomEvents_ExpectedSeconds=10
+	}
+
+function fillNumberOfTests
+	{
+		Device_ExpectedTests=113
+		Domoticz_ExpectedTests=71
+		EventHelpers_ExpectedTests=32
+		EventHelpersStorage_ExpectedTests=50
+		HTTPResponse_ExpectedTests=6
+		Lodash_ExpectedTests=100
+		ScriptdzVentsDispatching_ExpectedTests=2
+		TimedCommand_ExpectedTests=46
+		Time_ExpectedTests=340
+		Utils_ExpectedTests=25
+		Variable_ExpectedTests=15
+		ContactDoorLockInvertedSwitch_ExpectedTests=2
+		DelayedVariableScene_ExpectedTests=2
+		EventState_ExpectedTests=2
+		Integration_ExpectedTests=217
+		SelectorSwitch_ExpectedTests=2
+		SystemAndCustomEvents_ExpectedTests=7
 	}
 
 function testDir
 	{
-		underScore="_ExpectedSeconds"
+		underScoreSeconds="_ExpectedSeconds"
+		underScoreTests="_ExpectedTests"
 		testfilelist=(`ls ${prefix}*.lua`)
 		outfile=collectOuput
 		for i in ${testfilelist[@]}
 			do
 			testFile=${i%"$suffix"}
 			testFile=${testFile#"$prefix"}
-			Expected=$(( $(($testFile$underScore)) / factor ))
+			Expected=$(( $(($testFile$underScoreSeconds)) / factor ))
+			Tests=$(( $(($testFile$underScoreTests)) / 1 ))
+			if test $i == 'testSystemAndCustomEvents.lua' ; then
+				tput sc # save cursor
+				stopBackgroundProcesses 
+				sleep 5
+				tput rc;tput el # rc
+			fi
 			echo -n $(showTime) " "
-			printf "%-42s %4d  "  $testFile $Expected
+			printf "%-42s %4d "  $testFile $Expected 
+			printf "%10d" $Tests
 			# busted $i  2>&1 >/dev/null
 			busted $i  > $outfile
+			
 			if [[ $? -ne 0 ]];then
 				printf "%-10s" "===>> NOT OK"
 				echo
@@ -149,7 +191,7 @@ function testDir
 				timePassed=$(echo $bottomLine |awk '{print $13}')
 				totalTest=$((totalTest + succes))
 				faulty=$((failed + error))
-				printf "%-12s %*s %*s %11.4f" "      OK" 7 $succes 9 $faulty  $timePassed
+				printf "%-12s %*s %*s %11.4f" "	  OK" 7 $succes 9 $faulty  $timePassed
 			fi
 			rm $outfile
 			echo
@@ -160,26 +202,33 @@ stopBackgroundProcesses
 pstartSeconds=$SECONDS
 prefix="test"
 suffix=".lua"
-factor=10               # performance factor (about 10 on Raspberry)
+factor=10			   # performance factor (about 10 on Raspberry)
 
 cd $basedir/dzVents/runtime/integration-tests
-npm --silent start 2>&1 >/dev/null &
+export NODE_NO_WARNINGS=1
+npm --silent start --scripts-prepend-node-path 2>&1 >/dev/null &
 checkStarted "server" 10
 
 # Just to be sure we do not destroy something important without a backup
 cp $basedir/domoticz.db $basedir/domoticz.db_$$
 
+# Make sure we start with a clean sheet
+rm -f $basedir/domoticz.db
+
 cd $basedir
-./domoticz > domoticz.log$$ &
+cp dzVents/runtime/integration-tests/scriptTestCustomAndSystemEventsScript.lua scripts/dzVents/scripts/scriptTestCustomAndSystemEventsScript.lua
+
+cd $basedir
+./domoticz  -www 8080 -sslwww 444 > domoticz.log$$ &
 checkStarted "domoticz" 20
 
 clear
-echo "======================== $NewVersion ================+===================== Tests =======================+"
-printf "|%6s %11s %80s " "  time |" " test-script"  "| expected | result | successful | failed | seconds |"
-echo
-echo "==================================================+===================================================+"
+echo "========= domoticz $NewVersion, dzVents V$dzVersion =======+========================== Tests ==============================+"
+echo "  time  | test-script	                           | expected | tests | result  | successful |  failed  | seconds  |"
+echo "===================================================+===============================================================+"
 
 fillTimes
+fillNumberOfTests
 totalTest=0
 cd $basedir/dzVents/runtime/tests
 testDir
@@ -189,22 +238,21 @@ testDir
 echo
 echo
 
-
 cd $basedir
-expectedErrorCount=4
+expectedErrorCount=6
 grep "Results stage 2: SUCCEEDED" domoticz.log$$ 2>&1 >/dev/null
 if [[ $? -eq 0 ]];then
 	grep "Results stage 1: SUCCEEDED" domoticz.log$$ 2>&1 >/dev/null
 	if [[ $? -eq 0 ]];then
 		#echo Stage 1 and stage 2 of integration test Succeeded
-		errorCount=$(grep "Error" domoticz.log$$ | wc -l)
-		if [[ $errorCount -eq $expectedErrorCount ]];then
+		errorCount=$(grep "Error" domoticz.log$$ | grep -v CheckAuthToken | wc -l)
+		if [[ $errorCount -le $expectedErrorCount ]];then
 			#echo Errors are to be expected
 			echo -n
 		else
-			grep -i Error domoticz.log$$
+			grep -i Error  domoticz.log$$ | grep -v CheckAuthToken
 			stopBackgroundProcesses 1
-		fi		
+		fi
 	else
 		echo Stage 1 of integration test was not completely succesfull
 		grep -i fail domoticz.log$$
@@ -219,6 +267,6 @@ else
 fi
 
 echo Total tests: $totalTest
-echo Finished without erors after $(showTime)
+echo $(date)';' dzVents version $dzVersion tested without erors after $(showTime)
 stopBackgroundProcesses 0
 cleanup
